requirements.txt
# Math and Physics
scipy
pandas
llvmlite==0.41.1  # require by numba
numba==0.58.1
PyGLM==2.7.1        # Required for running examplesrendering as well
warp-lang==0.10.1
pybullet

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Core modules
moderngl==5.7.4
glfw==2.5.5
numpy

# Geometry Libraries
trimesh==4.0.5  # 3D geometry operations (requires scipy)
triangle==20230923 # 2D triangulation

# General IO
PyYAML==6.0  # Consider this over omegaconf==2.3.0?

# 3D Loading models
pywavefront==1.3.3  # OBJ only
#pygltflib==1.15.3   # GLTF and GLB

# Image loading and processing
Pillow==10.0.0
opencv-python==4.9.0.80

# Video Loading
av==10.0.0
imageio-ffmpeg

# Font rendering
freetype-py==2.3.0

# ImGUI
imgui==2.0.0
PyOpenGL==3.1.7

# Code Checking
pytest==7.3.1
flake8==6.0.0

# Data
h5py==3.10.0

# [Debugging and Profiling]
matplotlib==3.6.2
#CProfileV==1.0.7
scalene==1.5.31.1
